删除用户功能

+++ metoo/src/main/java/com/metoo/service/impl/UserServiceImpl.java
@@ -1,8 +1,11 @@
 package com.metoo.service.impl;
 
+import com.metoo.core.domain.order.OrderRepository;
 import com.metoo.core.domain.user.User;
 import com.metoo.core.domain.user.UserRepository;
 import com.metoo.dto.user.UserDTO;
+import com.metoo.exception.ErrorMap;
+import com.metoo.exception.MetooFormException;
 import com.metoo.service.UserService;
 import com.metoo.utils.MD5Utils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,6 +22,8 @@ public class UserServiceImpl implements UserService {
 
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private OrderRepository orderRepository;
 
     @Override
     public List<UserDTO> loadUsers() {
@@ -58,6 +63,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void removeUserById(Long id) {
+        Integer used = orderRepository.checkUserUsed(id);
+        if (used > 0) {
+            throw new MetooFormException(ErrorMap.ALREADY_IN_USE_USER);
+        }
         userRepository.delete(id);
     }
